% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations_file.R
\name{get_stations_file}
\alias{get_stations_file}
\title{Get station metadata and ids from local downloaded file with all automatic stations}
\usage{
get_stations_file()
}
\description{
Get station metadata and ids from local downloaded file with all automatic stations
}
