% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations_file.R
\name{get_stations_meteo}
\alias{get_stations_meteo}
\title{Get meteo station metadata and ids from local downloaded file with all automatic stations}
\usage{
get_stations_meteo()
}
\description{
Get meteo station metadata and ids from local downloaded file with all automatic stations
}
