% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations_file.R
\name{get_stations_pluvio}
\alias{get_stations_pluvio}
\title{Get pluvio station metadata and ids from local downloaded file with all automatic stations}
\usage{
get_stations_pluvio()
}
\description{
Get pluvio station metadata and ids from local downloaded file with all automatic stations
}
