% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stations_web.R
\name{get_stations_web}
\alias{get_stations_web}
\title{Get station metadata from the SNIRH
Currently not working!}
\usage{
get_stations_web()
}
\description{
Get station metadata from the SNIRH
Currently not working!
}
